refactor(server): extract recceiver/cf model and config
Also fixes the hardcoded "iocid" string literal to CFPropertyName.IOC_ID.
